+++ CHANGELOG.markdown
@@ -6,6 +6,9 @@
 ## Bugs fixed
 * Enable 'package' native type (issue #102)
 
+## Changes
+* Even in Permissive mode, don't resolve unknown variable (see #103)
+
 # v1.1.0 (2015/03/11)
 
 Critical bugs have been fixed, upgrade recommended.