+++ CHANGELOG.markdown
@@ -9,6 +9,7 @@
 ## Changes
 * Even in Permissive mode, don't resolve unknown variable (see #103)
 * Add priority to the logger permissive output (see #106)
+* New hruby version
 
 ## Various
 * Hiera config interpolation logs decrease from NOTICE to INFO